# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "Product API with Auth + JWT",
        "contact": {
            "email": "hadiatab16@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Product"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register Account",
                "description": "Create new account for accessing data",
                "operationId": "registration",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "test@email.com",
                                    "password": "test1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "User created successfully",
                                    "user": {
                                        "name": "hab",
                                        "email": "hab@email.com",
                                        "updated_at": "2024-08-10T09:54:57.000000Z",
                                        "created_at": "2024-08-10T09:54:57.000000Z",
                                        "id": 1
                                    },
                                    "authorization": {
                                        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL3JlZ2lzdGVyIiwiaWF0IjoxNzIzMjgzNjk3LCJleHAiOjE3MjMyODcyOTcsIm5iZiI6MTcyMzI4MzY5NywianRpIjoiTXpKdjQ4b2RkRGpFSFM4RCIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.UcjMqKzsI58tcHvtC23s6fGD8xIrqTTCMUEtrg9RnsM",
                                        "type": "bearer"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login Account",
                "description": "Login to account",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "test@email.com",
                                    "password": "test1234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "User created successfully",
                                    "user": {
                                        "name": "hab",
                                        "email": "hab@email.com",
                                        "updated_at": "2024-08-10T09:54:57.000000Z",
                                        "created_at": "2024-08-10T09:54:57.000000Z",
                                        "id": 1
                                    },
                                    "authorization": {
                                        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL3JlZ2lzdGVyIiwiaWF0IjoxNzIzMjgzNjk3LCJleHAiOjE3MjMyODcyOTcsIm5iZiI6MTcyMzI4MzY5NywianRpIjoiTXpKdjQ4b2RkRGpFSFM4RCIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.UcjMqKzsI58tcHvtC23s6fGD8xIrqTTCMUEtrg9RnsM",
                                        "type": "bearer"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Info",
                "description": "Get detail user",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "User created successfully",
                                    "user": {
                                        "name": "hab",
                                        "email": "hab@email.com",
                                        "updated_at": "2024-08-10T09:54:57.000000Z",
                                        "created_at": "2024-08-10T09:54:57.000000Z",
                                        "id": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log the user out",
                "description": "Invalidate the token.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Successfully logged out"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/refresh": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh the token",
                "description": "Generate a new token.",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "Token refreshed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token to access these endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}